refactor: common initial_provide logfinder
